(db delete required) use Flyway instead of hbm2ddl for db migrations going forward. rename "device id" as "detection id". add last state change for device connected.

+++ src/main/java/net/forlevity/homeglue/api/ApplianceDetectorDto.java
@@ -24,7 +24,7 @@ import java.time.Instant;
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class ApplianceDetectorDto {
 
-    Float minWatts;
+    Double minWatts;
     Integer offDelaySeconds;
     Integer maxOnSeconds;
     Boolean on;
